// src/common/enums.ts
export enum NodeType {
  ATTRIBUTE_NODE = 1,
  CDATA_SECTION_NODE = 2,
  COMMENT_NODE = 3,
  DOCUMENT_FRAGMENT_NODE = 4,
  DOCUMENT_NODE = 5,
  DOCUMENT_TYPE_NODE = 6,
  ELEMENT_NODE = 7,
  ENTITY_NODE = 8,
  ENTITY_REFERENCE_NODE = 9,
  NOTATION_NODE = 10,
  PROCESSING_INSTRUCTION_NODE = 11,
  TEXT_NODE = 12,
}